FPII-2128:Denial of Service Vulnerability in Qualcomm Bootloader (Device Specific) CVE-2015-8893 ANDROID-28822690
A denial of service vulnerability in the Qualcomm bootloader could enable a local malicious application to cause a local permanent device compromise, which may require reflashing the operating system to repair the device. This issue is rated as Moderate because it first requires compromising a privileged process.
CVE-2015-8893
ANDROID-28822690
CR#822275

Change-Id: I6ce2fa39074ab17ab82df35495902633ad4c0fab